Created ConfirmationClinicPopUp and ConfirmationServicePopUp and added max input length
