+++ app/src/main/res/layout/fragment_create_clinic.xml
@@ -85,6 +85,7 @@
                 android:background="@drawable/round_style"
                 android:backgroundTint="@color/white"
                 android:hint="@string/NameHint"
+                android:maxLength="20"
                 android:inputType="textNoSuggestions"
                 android:textColorHint="@color/input_grey"
                 android:textSize="16sp" />
@@ -117,6 +118,7 @@
                 android:background="@drawable/round_style"
                 android:backgroundTint="@color/white"
                 android:hint="@string/LastNameHint"
+                android:maxLength="20"
                 android:inputType="textNoSuggestions"
                 android:textColor="@color/black"
                 android:textColorHint="@color/input_grey"
@@ -148,6 +150,7 @@
                 android:background="@drawable/round_style"
                 android:backgroundTint="@color/white"
                 android:hint="@string/AfmHint"
+                android:maxLength="9"
                 android:inputType="number"
                 android:textColor="@color/black"
                 android:textColorHint="@color/input_grey"
@@ -180,6 +183,7 @@
                 android:background="@drawable/round_style"
                 android:backgroundTint="@color/white"
                 android:hint="@string/PhoneHint"
+                android:maxLength="10"
                 android:inputType="number"
                 android:textColor="@color/black"
                 android:textColorHint="@color/input_grey"
@@ -211,6 +215,7 @@
                 android:background="@drawable/round_style"
                 android:backgroundTint="@color/white"
                 android:hint="@string/NameHint"
+                android:maxLength="20"
                 android:inputType="textNoSuggestions"
                 android:textColor="@color/black"
                 android:textColorHint="@color/input_grey"
@@ -243,6 +248,7 @@
                 android:background="@drawable/round_style"
                 android:backgroundTint="@color/white"
                 android:hint="@string/CityHint"
+                android:maxLength="20"
                 android:inputType="textNoSuggestions"
                 android:textColor="@color/black"
                 android:textColorHint="@color/input_grey"
@@ -276,6 +282,7 @@
                 android:background="@drawable/round_style"
                 android:backgroundTint="@color/white"
                 android:hint="@string/AddressHint"
+                android:maxLength="20"
                 android:inputType="textNoSuggestions"
                 android:textColor="@color/black"
                 android:textColorHint="@color/input_grey"
@@ -308,6 +315,7 @@
                 android:background="@drawable/round_style"
                 android:backgroundTint="@color/white"
                 android:hint="@string/hint_username"
+                android:maxLength="20"
                 android:inputType="textNoSuggestions"
                 android:textColor="@color/black"
                 android:textColorHint="@color/input_grey" />
@@ -339,6 +347,7 @@
                 android:background="@drawable/round_style"
                 android:backgroundTint="@color/white"
                 android:hint="@string/hint_password"
+                android:maxLength="20"
                 android:inputType="textNoSuggestions"
                 android:textColor="@color/black"
                 android:textColorHint="@color/input_grey" />
