+++ app/src/main/res/layout/fragment_create_service.xml
@@ -40,6 +40,7 @@
                 android:backgroundTint="@color/white"
                 android:gravity="top"
                 android:hint="@string/DescriptionHint"
+                android:maxLength="70"
                 android:inputType="textMultiLine|textNoSuggestions"
                 android:textColor="@color/black"
                 android:textColorHint="@color/input_grey"
@@ -87,6 +88,7 @@
                 android:background="@drawable/round_style"
                 android:backgroundTint="@color/white"
                 android:hint="@string/NameHint"
+                android:maxLength="20"
                 android:inputType="textNoSuggestions"
                 android:textColor="@color/black"
                 android:textColorHint="@color/input_grey"
@@ -102,7 +104,7 @@
             android:layout_marginEnd="1dp"
             app:boxStrokeWidth="0dp"
             app:boxStrokeWidthFocused="0dp"
-            app:counterMaxLength="30"
+            app:counterMaxLength="20"
             app:endIconMode="clear_text"
             app:helperText="Υποχρεωτικό *"
             app:errorEnabled="true"
@@ -119,6 +121,7 @@
                 android:background="@drawable/round_style"
                 android:backgroundTint="@color/white"
                 android:hint="@string/codeString"
+                android:maxLength="20"
                 android:inputType="textNoSuggestions"
                 android:textColor="@color/black"
                 android:textColorHint="@color/input_grey"
@@ -134,7 +137,7 @@
             android:layout_marginEnd="1dp"
             app:boxStrokeWidth="0dp"
             app:boxStrokeWidthFocused="0dp"
-            app:counterMaxLength="10"
+            app:counterMaxLength="7"
             app:endIconMode="clear_text"
             app:errorEnabled="true"
             app:helperText="Υποχρεωτικό *"
@@ -151,6 +154,7 @@
                 android:background="@drawable/round_style"
                 android:backgroundTint="@color/white"
                 android:hint="Τιμή*"
+                android:maxLength="7"
                 android:inputType="number"
                 android:textColor="@color/black"
                 android:textColorHint="@color/input_grey"
